script: Move Trusted Types interfaces to `script/dom/trustedtypes`
Part of #38901
