+++ components/script/dom/mod.rs
@@ -414,11 +414,8 @@ pub(crate) mod touchevent;
 pub(crate) mod touchlist;
 pub(crate) mod transitionevent;
 pub(crate) mod treewalker;
-pub(crate) mod trustedhtml;
-pub(crate) mod trustedscript;
-pub(crate) mod trustedscripturl;
-pub(crate) mod trustedtypepolicy;
-pub(crate) mod trustedtypepolicyfactory;
+pub(crate) mod trustedtypes;
+pub(crate) use self::trustedtypes::*;
 pub(crate) mod uievent;
 pub(crate) mod url;
 pub(crate) mod urlhelper;
