+++ components/script/dom/trustedtypes/trustedscript.rs
@@ -15,8 +15,10 @@ use crate::dom::bindings::root::DomRoot;
 use crate::dom::bindings::str::DOMString;
 use crate::dom::csp::CspReporting;
 use crate::dom::globalscope::GlobalScope;
-use crate::dom::trustedtypepolicy::TrustedType;
-use crate::dom::trustedtypepolicyfactory::{DEFAULT_SCRIPT_SINK_GROUP, TrustedTypePolicyFactory};
+use crate::dom::trustedtypes::trustedtypepolicy::TrustedType;
+use crate::dom::trustedtypes::trustedtypepolicyfactory::{
+    DEFAULT_SCRIPT_SINK_GROUP, TrustedTypePolicyFactory,
+};
 use crate::script_runtime::{CanGc, JSContext};
 
 #[dom_struct]
