+++ components/script/dom/trustedtypes/trustedscripturl.rs
@@ -6,28 +6,27 @@ use std::fmt;
 
 use dom_struct::dom_struct;
 
-use crate::conversions::Convert;
-use crate::dom::bindings::codegen::Bindings::TrustedHTMLBinding::TrustedHTMLMethods;
-use crate::dom::bindings::codegen::UnionTypes::{
-    TrustedHTMLOrNullIsEmptyString, TrustedHTMLOrString,
-};
+use crate::dom::bindings::codegen::Bindings::TrustedScriptURLBinding::TrustedScriptURLMethods;
+use crate::dom::bindings::codegen::UnionTypes::TrustedScriptURLOrUSVString;
 use crate::dom::bindings::error::Fallible;
 use crate::dom::bindings::reflector::{Reflector, reflect_dom_object};
 use crate::dom::bindings::root::DomRoot;
 use crate::dom::bindings::str::DOMString;
 use crate::dom::globalscope::GlobalScope;
-use crate::dom::trustedtypepolicy::TrustedType;
-use crate::dom::trustedtypepolicyfactory::{DEFAULT_SCRIPT_SINK_GROUP, TrustedTypePolicyFactory};
+use crate::dom::trustedtypes::trustedtypepolicy::TrustedType;
+use crate::dom::trustedtypes::trustedtypepolicyfactory::{
+    DEFAULT_SCRIPT_SINK_GROUP, TrustedTypePolicyFactory,
+};
 use crate::script_runtime::CanGc;
 
 #[dom_struct]
-pub struct TrustedHTML {
+pub struct TrustedScriptURL {
     reflector_: Reflector,
 
     data: DOMString,
 }
 
-impl TrustedHTML {
+impl TrustedScriptURL {
     fn new_inherited(data: DOMString) -> Self {
         Self {
             reflector_: Reflector::new(),
@@ -39,25 +38,28 @@ impl TrustedHTML {
         reflect_dom_object(Box::new(Self::new_inherited(data)), global, can_gc)
     }
 
-    pub(crate) fn get_trusted_script_compliant_string(
+    pub(crate) fn get_trusted_script_url_compliant_string(
         global: &GlobalScope,
-        value: TrustedHTMLOrString,
-        sink: &str,
+        value: TrustedScriptURLOrUSVString,
+        containing_class: &str,
+        field: &str,
         can_gc: CanGc,
     ) -> Fallible<DOMString> {
         match value {
-            TrustedHTMLOrString::String(value) => {
+            TrustedScriptURLOrUSVString::USVString(value) => {
+                let sink = format!("{} {}", containing_class, field);
                 TrustedTypePolicyFactory::get_trusted_type_compliant_string(
-                    TrustedType::TrustedHTML,
+                    TrustedType::TrustedScriptURL,
                     global,
-                    value,
-                    sink,
+                    value.as_ref().into(),
+                    &sink,
                     DEFAULT_SCRIPT_SINK_GROUP,
                     can_gc,
                 )
             },
-
-            TrustedHTMLOrString::TrustedHTML(trusted_html) => Ok(trusted_html.data.clone()),
+            TrustedScriptURLOrUSVString::TrustedScriptURL(trusted_script_url) => {
+                Ok(trusted_script_url.data.clone())
+            },
         }
     }
 
@@ -66,34 +68,21 @@ impl TrustedHTML {
     }
 }
 
-impl fmt::Display for TrustedHTML {
+impl fmt::Display for TrustedScriptURL {
     #[inline]
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         f.write_str(&self.data.str())
     }
 }
 
-impl TrustedHTMLMethods<crate::DomTypeHolder> for TrustedHTML {
-    /// <https://www.w3.org/TR/trusted-types/#trustedhtml-stringification-behavior>
+impl TrustedScriptURLMethods<crate::DomTypeHolder> for TrustedScriptURL {
+    /// <https://www.w3.org/TR/trusted-types/#trustedscripturl-stringification-behavior>
     fn Stringifier(&self) -> DOMString {
         self.data.clone()
     }
 
-    /// <https://www.w3.org/TR/trusted-types/#dom-trustedhtml-tojson>
+    /// <https://www.w3.org/TR/trusted-types/#dom-trustedscripturl-tojson>
     fn ToJSON(&self) -> DOMString {
         self.data.clone()
-    }
-}
-
-impl Convert<TrustedHTMLOrString> for TrustedHTMLOrNullIsEmptyString {
-    fn convert(self) -> TrustedHTMLOrString {
-        match self {
-            TrustedHTMLOrNullIsEmptyString::TrustedHTML(trusted_html) => {
-                TrustedHTMLOrString::TrustedHTML(trusted_html)
-            },
-            TrustedHTMLOrNullIsEmptyString::NullIsEmptyString(str) => {
-                TrustedHTMLOrString::String(str)
-            },
-        }
     }
 }
