+++ components/script/dom/trustedtypes/trustedtypepolicyfactory.rs
@@ -22,10 +22,10 @@ use crate::dom::bindings::str::DOMString;
 use crate::dom::csp::CspReporting;
 use crate::dom::eventtarget::EventTarget;
 use crate::dom::globalscope::GlobalScope;
-use crate::dom::trustedhtml::TrustedHTML;
-use crate::dom::trustedscript::TrustedScript;
-use crate::dom::trustedscripturl::TrustedScriptURL;
-use crate::dom::trustedtypepolicy::{TrustedType, TrustedTypePolicy};
+use crate::dom::trustedtypes::trustedhtml::TrustedHTML;
+use crate::dom::trustedtypes::trustedscript::TrustedScript;
+use crate::dom::trustedtypes::trustedscripturl::TrustedScriptURL;
+use crate::dom::trustedtypes::trustedtypepolicy::{TrustedType, TrustedTypePolicy};
 use crate::script_runtime::{CanGc, JSContext};
 
 #[dom_struct]
